+++ src/parser.ts
@@ -8,7 +8,7 @@ import { tuple } from 'fp-ts/lib/function'
 
 const isDigit = (c: string): boolean => '0123456789'.indexOf(c) !== -1
 
-const isPunctuation = (c: string): boolean => '| =\n():,{};[]'.indexOf(c) !== -1
+const isPunctuation = (c: string): boolean => '| =\n():,{};[]->'.indexOf(c) !== -1
 
 const identifierFirstLetter = P.sat(c => !isDigit(c) && !isPunctuation(c))
 
@@ -22,19 +22,54 @@ export const identifier: P.Parser<string> = expected(
   P.fold([identifierFirstLetter, C.many(identifierBody)])
 )
 
-const singletonType: P.Parser<M.Type> = identifier.map(name => M.type(name))
+const leftParens: P.Parser<string> = P.fold([C.char('('), S.spaces])
 
-const unparenthesizedType: P.Parser<M.Type> = identifier.chain(name =>
-  S.spaces.applySecond(P.many(type).map(types => M.type(name, types)))
+const rightParens: P.Parser<string> = P.fold([S.spaces, C.char(')')])
+
+const withParens = <A>(parser: P.Parser<A>): P.Parser<A> => {
+  return leftParens.applySecond(parser).applyFirst(rightParens)
+}
+
+const typeReferenceLeaf: P.Parser<M.Type> = identifier.map(name => M.typeReference(name))
+
+export const typeReference: P.Parser<M.Type> = identifier.chain(name =>
+  S.spaces.applySecond(types.map(parameters => M.typeReference(name, parameters)))
 )
-
-const parenthesizedType: P.Parser<M.Type> = P.fold([C.char('('), S.spaces]).applySecond(
-  unparenthesizedType.applyFirst(P.fold([S.spaces, C.char(')')]))
-)
-
-export const type: P.Parser<M.Type> = expected('a type', parenthesizedType.alt(unparenthesizedType))
 
 const comma = P.fold([S.spaces, C.char(','), S.spaces])
+
+const other: P.Parser<Array<M.Type>> = comma.chain(() => P.sepBy(comma, type)).alt(P.parser.of([]))
+
+export const tupleType: P.Parser<M.Type> = expected(
+  'a tuple',
+  leftParens
+    .chain(() =>
+      type.applyFirst(comma).chain(fst => type.chain(snd => other.map(other => M.tupleType(fst, snd, other))))
+    )
+    .applyFirst(rightParens)
+)
+
+const arrow = P.fold([S.spaces, S.string('->'), S.spaces])
+
+export const functionType: P.Parser<M.Type> = expected(
+  'a function type',
+  S.spaces.chain(() =>
+    withParens(functionType)
+      .alt(typeReference)
+      .alt(tupleType)
+      .chain(domain => arrow.applySecond(type).map(codomain => M.functionType(domain, codomain)))
+  )
+)
+
+export const type: P.Parser<M.Type> = functionType.alt(typeReference).alt(tupleType)
+
+export const types: P.Parser<Array<M.Type>> = P.sepBy(
+  S.spaces,
+  functionType
+    .alt(typeReferenceLeaf)
+    .alt(withParens(typeReference))
+    .alt(tupleType)
+)
 
 const pair: P.Parser<[string, M.Type]> = identifier.chain(name =>
   P.fold([S.spaces, S.string('::'), S.spaces])
@@ -42,19 +77,21 @@ const pair: P.Parser<[string, M.Type]> = identifier.chain(name =>
     .map(type => tuple(name, type))
 )
 
-const recordMembers: P.Parser<Array<M.Member>> = P.fold([C.char('{'), S.spaces])
-  .applySecond(P.sepBy(comma, pair).map(ps => ps.map(([name, type]) => M.member(type, some(name)))))
+const objectType: P.Parser<Array<[string, M.Type]>> = P.fold([C.char('{'), S.spaces])
+  .applySecond(P.sepBy(comma, pair))
   .applyFirst(P.fold([S.spaces, C.char('}')]))
 
-const positionalMember: P.Parser<M.Member> = parenthesizedType.alt(singletonType).map(type => M.member(type))
-
-const positionalMembers: P.Parser<Array<M.Member>> = P.sepBy(S.spaces, positionalMember)
-
-const members = recordMembers.alt(positionalMembers).alt(P.failWith('invalid constructor members'))
-
-export const constructor: P.Parser<M.Constructor> = identifier.chain(name =>
-  S.spaces.applySecond(members.map(members => M.constructor(name, members)))
+const recordConstructor: P.Parser<M.Constructor> = identifier.chain(name =>
+  S.spaces.applySecond(
+    objectType.map(pairs => M.constructor(name, pairs.map(([name, type]) => M.member(type, some(name)))))
+  )
 )
+
+const positionalConstructor: P.Parser<M.Constructor> = identifier.chain(name =>
+  S.spaces.applySecond(types.map(types => M.constructor(name, types.map(type => M.member(type)))))
+)
+
+export const constructor: P.Parser<M.Constructor> = recordConstructor.alt(positionalConstructor)
 
 const equal = P.fold([S.spaces, C.char('='), S.spaces])