+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 > **Tags:**
+>
 > - [New Feature]
 > - [Bug Fix]
 > - [Breaking Change]
@@ -12,3 +13,14 @@
 **Note**: Gaps between patch versions are faulty/broken releases.
 **Note**: A feature tagged as Experimental is in a high state of flux, you're at risk of it changing without notice.
 
+# 0.0.2
+
+- **Breaking Change**
+  - `model.ts` is changed in order to support tuple types and function types (@gcanti)
+- **New Feature**
+  - add support for tuple types (@gcanti)
+  - add support for function types (@gcanti)
+
+# 0.0.1
+
+Initial release